moved moved wayland directory out of frontend and renamed it frontend_wayland, also moved protocol directory into it
