+++ src/server/frontend_wayland/null_event_sink.h
