+++ src/server/frontend_wayland/output_manager.h
