+++ src/server/frontend_wayland/protocol/wrapper_generator.cpp
